# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BHRM Teams API
  version: 2.0.0
  description: |
    Enterprise Task-Centric Execution Platform — REST API.

    ## Authentication
    All protected routes require `Authorization: Bearer <token>` obtained from
    `POST /auth/login` or `POST /auth/refresh`.

    ## Response envelopes
    | Type | Shape |
    |---|---|
    | Single resource | `{ "success": true, "data": { … } }` |
    | Collection | `{ "success": true, "data": [ … ], "meta": { total, page, limit, totalPages } }` |
    | Error | `{ "success": false, "error": { "code": "…", "message": "…", "statusCode": 4xx } }` |

    ## Role model
    | Role | Description |
    |---|---|
    | `admin` | Company owner — full access to all company resources |
    | `staff` | Invited member — access scoped to assigned tasks |

    ## Endpoint naming conventions
    - Resource-based URLs, nouns only
    - State transitions via `PATCH /{resource}/{id}/status`
    - Bulk assignment via `PATCH /{resource}/{id}/assignees`
    - Avoid custom action verbs in URLs
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# COMPONENTS
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    StaffStatus:
      type: string
      enum: [invited, active, suspended]
...
    SprintStatus:
      type: string
      enum: [draft, active, completed]
...
    PaginatedMeta:
      type: object
      properties:
        total:      { type: integer, example: 42 }
        page:       { type: integer, example: 1 }
        limit:      { type: integer, example: 20 }
        totalPages: { type: integer, example: 3 }

    # ── Error ─────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code:       { type: string, example: NOT_FOUND }
            message:    { type: string, example: Resource not found }
            statusCode: { type: integer, example: 404 }

    # ── User ─────────────────────────────────────────────────────────────────
    UserSummary:
      type: object
      properties:
        id:        { type: string, format: uuid }
        fullName:  { type: string, example: Jane Smith }
        email:     { type: string, format: email }
        avatarUrl: { type: string, format: uri, nullable: true }
...
    StaffMember:
      type: object
      properties:
        id:        { type: string, format: uuid }
        user:      { $ref: '#/components/schemas/UserSummary' }
        companyId: { type: string, format: uuid }
        status:    { $ref: '#/components/schemas/StaffStatus' }
        joinedAt:  { type: string, format: date-time, nullable: true }
        invitedAt: { type: string, format: date-time }
...
    Sprint:
      type: object
      properties:
        id:        { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        name:      { type: string, example: Sprint 1 }
        goal:      { type: string, nullable: true }
        status:    { $ref: '#/components/schemas/SprintStatus' }
        startDate: { type: string, format: date, nullable: true }
        endDate:   { type: string, format: date, nullable: true }
        taskCount: { type: integer }
        createdAt: { type: string, format: date-time }
...
    TaskStep:
      type: object
      properties:
        id:        { type: string, format: uuid }
        taskId:    { type: string, format: uuid }
        title:     { type: string }
        isChecked: { type: boolean }
        order:     { type: integer }
        checkedAt: { type: string, format: date-time, nullable: true }
        checkedBy: { $ref: '#/components/schemas/UserSummary', nullable: true }
        createdAt: { type: string, format: date-time }

    # ── Attachment ────────────────────────────────────────────────────────────
...
  responses:
...
    Forbidden:
      description: Insufficient role or resource ownership
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    MessageOk:
      description: Operation succeeded
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean }
              data:
                type: object
                properties:
                  message: { type: string }

  # ── Reusable parameters ────────────────────────────────────────────────────
  parameters:
    companyId:    { name: companyId,    in: path,  required: true, schema: { type: string, format: uuid } }
    projectId:    { name: projectId,    in: path,  required: true, schema: { type: string, format: uuid } }
    sprintId:     { name: sprintId,     in: path,  required: true, schema: { type: string, format: uuid } }
    taskId:       { name: taskId,       in: path,  required: true, schema: { type: string, format: uuid } }
    staffId:      { name: staffId,      in: path,  required: true, schema: { type: string, format: uuid } }
...
    stepId:       { name: stepId,       in: path,  required: true, schema: { type: string, format: uuid } }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    search:
      name: search
      in: query
      schema: { type: string }

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /companies/{companyId}/staff:
    get:
      tags: [Staff]
      summary: List staff members (admin only)
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/StaffStatus' }
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: Paginated staff list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/StaffMember' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /companies/{companyId}/staff/{staffId}:
    get:
      tags: [Staff]
      summary: Get individual staff member profile (admin only)
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/staffId'
      responses:
        '200':
          description: Staff member
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/StaffMember' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Staff]
      summary: Update staff member metadata (admin only)
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/staffId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { $ref: '#/components/schemas/StaffStatus' }
      responses:
        '200':
          description: Updated staff member
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/StaffMember' }
        '400': { description: Invalid status transition }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Staff]
      summary: Remove staff member (admin only; soft-delete)
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/staffId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /companies/{companyId}/staff/{staffId}/suspend:
    patch:
      tags: [Staff]
      summary: Suspend an active staff member (admin only)
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/staffId'
      responses:
        '200':
          description: Staff suspended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/StaffMember' }
        '400': { description: Staff is not active }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /companies/{companyId}/staff/{staffId}/activate:
    patch:
      tags: [Staff]
      summary: Reactivate a suspended staff member (admin only)
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/staffId'
      responses:
        '200':
          description: Staff activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/StaffMember' }
        '400': { description: Staff is not suspended }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /companies/{companyId}/staff/{staffId}/resend-invite:
    post:
      tags: [Staff]
      summary: Resend invite email (admin only; staff must be in invited status)
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/staffId'
      responses:
        '200':
          description: Invite resent with refreshed token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/StaffMember' }
        '400': { description: Staff is not in invited status }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /projects/{projectId}/sprints:
    get:
      tags: [Sprints]
      summary: List sprints for a project
      parameters:
        - $ref: '#/components/parameters/projectId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/SprintStatus' }
      responses:
        '200':
          description: Paginated sprint list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Sprint' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '404': { $ref: '#/components/responses/NotFound' }

    post:
      tags: [Sprints]
      summary: Create a sprint in draft status (admin only)
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:      { type: string }
                goal:      { type: string }
                startDate: { type: string, format: date }
                endDate:   { type: string, format: date }
      responses:
        '201':
          description: Sprint created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Sprint' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /sprints/{sprintId}:
    get:
      tags: [Sprints]
      summary: Get sprint details
      parameters:
        - $ref: '#/components/parameters/sprintId'
      responses:
        '200':
          description: Sprint
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Sprint' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      tags: [Sprints]
      summary: Update sprint metadata (admin only)
      parameters:
        - $ref: '#/components/parameters/sprintId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:      { type: string }
                goal:      { type: string }
                startDate: { type: string, format: date }
                endDate:   { type: string, format: date }
      responses:
        '200':
          description: Updated sprint
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/Sprint' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Sprints]
      summary: Delete a draft sprint (admin only)
      parameters:
        - $ref: '#/components/parameters/sprintId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '400': { description: Can only delete draft sprints }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /tasks/{taskId}/steps:
    get:
      tags: [Steps]
      summary: List steps for a task ordered by position
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Step list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/TaskStep' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Steps]
      summary: Add a step to a task (admin only)
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 500 }
                order: { type: integer, minimum: 0, description: 'Defaults to last position' }
      responses:
        '201':
          description: Step created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/TaskStep' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tasks/{taskId}/steps/reorder:
    patch:
      tags: [Steps]
      summary: Reorder all steps by providing the complete ordered list of IDs (admin only)
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stepIds]
              properties:
                stepIds:
                  type: array
                  description: Must include every active step ID for this task — no omissions
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: Updated step list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/TaskStep' }
        '400': { description: stepIds set does not match the task's active steps exactly }
        '403': { $ref: '#/components/responses/Forbidden' }
  /tasks/{taskId}/steps/{stepId}:
    patch:
      tags: [Steps]
      summary: |
        Update a step (assignee or admin).
        Use `{ "isChecked": true }` to complete, `{ "isChecked": false }` to reopen.
        Also supports updating `title`.
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/stepId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isChecked: { type: boolean }
                title:     { type: string, maxLength: 500 }
            examples:
              complete:
                summary: Mark complete
                value: { isChecked: true }
              reopen:
                summary: Mark incomplete
                value: { isChecked: false }
              rename:
                summary: Rename step
                value: { title: 'Updated step title' }
      responses:
        '200':
          description: Updated step
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/TaskStep' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Steps]
      summary: Delete a step (admin only; soft-delete)
      parameters:
        - $ref: '#/components/parameters/taskId'
        - $ref: '#/components/parameters/stepId'
      responses:
        '200': { $ref: '#/components/responses/MessageOk' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ════════════════════════════════════════════════════════════════════════════
  # ATTACHMENTS
  # ════════════════════════════════════════════════════════════════════════════
...